move class files to the correct package
